Update code-of-conduct.md
Added a statement relating to steering toward harmony with the community at large. Just a draft I put together. Open to any changes or improvements to it, or any idea why not to include it. One thought that I had was that it may cheapen the rest of the anti-harassment policy that aims to make diversity possible.